Added investigate a dataset project
In this investigation, we look at what factors determine the success of a movie. Some questions explored include:
Do big budgets guarantee a big return on Investments?
Does a star cast guarantee the movie will be a success?
Does a good production company have any bearing on the success of a movie?

The original dataset is provided should you want to make your own analysis however, the cleaned dataset is provided in the SQLite engine file attached should you want to reproduce the analysis.